Rebased from main and refactored code to fit new structure We used rebase commanmd to follow up on all new push. All files have been moved with git mv if they need to be moved and som files have been renamed. Some of the test needs to be rewriten
