+++ Chirp.sln
@@ -7,7 +7,15 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{8BDDB909-80D
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Razor", "src\Chirp.Razor\Chirp.Razor.csproj", "{AA5AD6CA-EF27-485D-B65C-4A7DF9B19752}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.CLI.Client.Tests", "test\Chirp.CLI.Client.Tests\Chirp.CLI.Client.Tests.csproj", "{49D65451-691E-4F0E-B7F6-506EE7C4C229}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Infrastructure", "src\Chirp.Infrastructure\Chirp.Infrastructure.csproj", "{F791ACE8-50D4-495D-8453-BE3823B53F91}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Core", "src\Chirp.Core\Chirp.Core.csproj", "{A96A7C62-13F3-42D6-A16D-9BE67EE31D01}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{4E187788-4FDE-4356-BBEF-779B60E19525}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Razor.Tests", "test\Chirp.Razor.Tests\Chirp.Razor.Tests.csproj", "{D402284E-9903-4F0B-9410-2C9C3CA05547}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Infrastructure.Tests", "test\Chirp.Infrastructure.Tests\Chirp.Infrastructure.Tests.csproj", "{80A8BDAE-CEEB-4AB9-B41E-59860C2AFF99}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,8 +30,28 @@ Global
 		{AA5AD6CA-EF27-485D-B65C-4A7DF9B19752}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AA5AD6CA-EF27-485D-B65C-4A7DF9B19752}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AA5AD6CA-EF27-485D-B65C-4A7DF9B19752}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F791ACE8-50D4-495D-8453-BE3823B53F91}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F791ACE8-50D4-495D-8453-BE3823B53F91}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F791ACE8-50D4-495D-8453-BE3823B53F91}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F791ACE8-50D4-495D-8453-BE3823B53F91}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A96A7C62-13F3-42D6-A16D-9BE67EE31D01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A96A7C62-13F3-42D6-A16D-9BE67EE31D01}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A96A7C62-13F3-42D6-A16D-9BE67EE31D01}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A96A7C62-13F3-42D6-A16D-9BE67EE31D01}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D402284E-9903-4F0B-9410-2C9C3CA05547}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D402284E-9903-4F0B-9410-2C9C3CA05547}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D402284E-9903-4F0B-9410-2C9C3CA05547}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D402284E-9903-4F0B-9410-2C9C3CA05547}.Release|Any CPU.Build.0 = Release|Any CPU
+		{80A8BDAE-CEEB-4AB9-B41E-59860C2AFF99}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{80A8BDAE-CEEB-4AB9-B41E-59860C2AFF99}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{80A8BDAE-CEEB-4AB9-B41E-59860C2AFF99}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{80A8BDAE-CEEB-4AB9-B41E-59860C2AFF99}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{AA5AD6CA-EF27-485D-B65C-4A7DF9B19752} = {8BDDB909-80D3-4E33-9978-218ED07502DE}
+		{F791ACE8-50D4-495D-8453-BE3823B53F91} = {8BDDB909-80D3-4E33-9978-218ED07502DE}
+		{A96A7C62-13F3-42D6-A16D-9BE67EE31D01} = {8BDDB909-80D3-4E33-9978-218ED07502DE}
+		{D402284E-9903-4F0B-9410-2C9C3CA05547} = {4E187788-4FDE-4356-BBEF-779B60E19525}
+		{80A8BDAE-CEEB-4AB9-B41E-59860C2AFF99} = {4E187788-4FDE-4356-BBEF-779B60E19525}
 	EndGlobalSection
 EndGlobal
