+++ src/Chirp.Infrastructure/Author.cs
@@ -1,0 +1,9 @@
+﻿namespace Chirp.Infrastructure;
+
+public class Author
+{
+    public int AuthorId { get; set; }
+    public required string Name { get; set; }
+    public required string Email { get; set; }
+    public required List<Cheep> Cheeps;
+}
