+++ src/Chirp.Infrastructure/Cheep.cs
@@ -1,0 +1,9 @@
+namespace Chirp.Infrastructure;
+
+public class Cheep
+{
+    public int CheepId { get; set; }
+    public required Author Author { get; set; }
+    public required string Text { get; set; }
+    public required DateTime TimeStamp { get; set; }
+}
