--- src/Chirp.Razor/CheepRecord.cs
@@ -1,4 +1,0 @@
-using System.Globalization;
-
-namespace CheepRecord;
-public record CheepViewModel (int? id,string? Author, string? Message, string? Timestamp);
