+++ src/Chirp.Razor/Chirp.Razor.csproj
@@ -26,4 +26,8 @@
     <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="7.0.10" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="../Chirp.Infrastructure/Chirp.Infrastructure.csproj" />
+  </ItemGroup>
+
 </Project>
