+++ src/Chirp.Razor/DbInitializer.cs
@@ -1,5 +1,5 @@
 using System;
-using CheepDB;
+using Chirp.Infrastructure;
 
 namespace Initializer;
