+++ src/Chirp.Razor/Program.cs
@@ -1,13 +1,10 @@
-using Chirp.Razor.Pages;
-using CheepDB;
-using Repository;
 using Initializer;
+using Chirp.Infrastructure;
 
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
 builder.Services.AddRazorPages();
-builder.Services.AddSingleton<ICheepService, CheepService>();
 
 
 /* Lecture notes for reference for later work. 
@@ -44,6 +41,8 @@ if (!app.Environment.IsDevelopment())
     // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
     app.UseHsts();
 }
+
+
 
 app.UseHttpsRedirection();
 app.UseStaticFiles();
