+++ test/Chirp.Infrastructure.Tests/GlobalUsings.cs
@@ -1,0 +1,4 @@
+global using Xunit;
+global using Chirp.Core;
+global using Chirp.Infrastructure;
+global using Chirp.Razor;
