--- test/Chirp.Razor.Test/GlobalUsings.cs
@@ -1,4 +1,0 @@
-global using Xunit;
-global using Chirp.Razor.Pages;
-global using CheepRecord;
-global using SQLDB;
